Add in the guts of the combat tracker
This needs a lot of testing and front end work.

It has a deck class that manages a deck of cards for one of the
two factions involved in a combat.

It has a hand class that manages a set of cards belonging to one
participant in a fight.

The cards module has meta information about a standard deck
including a mapping from index to symbol (two characters) to
name (a string).

The two main classes are Deadlands Classic overrides of combat
and combatant. Combat manages the two decks and each combatant
manages a hand of cards.