Make Card images configurable
Add a game setting to choose between the Deadlands cards and
the plainer "Modern" cards.

Generate the icon fields of the cards based on the choice.

Note: need to restart the application after changing this. It is a local
setting and each player can suit themselves, the setting is saved on
their machine, not in the global database. This also means you don't
need elevated permissions to set it for yourself.