Various bug fixes in dlc-combat
Add in a hasPreviousTurns getter.

Make sortCombatants a method. Instead of everywhere just doing
this.turns.sort directly,

Fix previous turn, so it sorts the combatants.

Fix the sorting of combatants in setupTurns. Was causing strange issues
when the round was started and a sleeved card was toggled in and out of
use.

Fix #sortCombatants, when sleeved cards and joker are toggled into use,
initiative becomes a number which is not a card. Explicitly checking for
initiative being a card was forcing them to the wrong end of the sort
order.