Improve shuffling behaviour
Combatant
=========
Change cleanupRound to return the cards cleaned up instead of discarding
them back to their deck of origin. Also, don't save parent state to
flags when called, parent will store state if necessary.

Combat
======
Stop storing the decks' shuffle flags to the Combat state flags. These
are really only used in resetRoundData and there seems to be a bug in
Foundry that was clearing them.

Merge resetRoundData and reconcileDeck.

Tidy up reapDiscards to not continually write combat's flags, once per
combatant.

Deck
====
Remove the setting of the shuffle flag while drawing. This is now only
ever set by reconcile, and is then immediately used by the combat in
resetRoundData.

Change discard to use a filter to eliminate duplicates from discards. It
no longer returns the number of cards discarded as nothing was using
it.