Fix some combat tracker bugs
Now properly handles end of turn when there are slleved cards or
jokers.

Some tweaks to the character sheet.

Add a configuration file, really belongs in character sheet, but
has a banner I've added.

Some language additions also really for character sheet, but doing
no harm in the combat tracker.

Change the icon for draw cards to the one I added for it, stop using
the d20.