Bugfix Sleeving not undoable
Sleeving a card was not storing enough state to allow it to be
undone.

While fixing that noticed that nextTurn, Vamoose and sleeveCard
had a lot of common code, refactored the common stuff. As part of
this changed the sleeveHightest socket operation to not take
a combatantId as only the acting combatant can sleeve a card.

Also noticed that the now extracted common saveTurn function
had a lot in common with previousTurn. Extracted the common
bits into another function.

Removed some redundant stuff in the socket implementation.