Tidy up cards
Make the Cards.mjs into CanonicalCards and put it in the global
scope.

Remove all references to 52 and 53 as specific jokers and
replace them with calls to get the red joker or black joker
index from CanonicalCards. This means that we now only have to
swap the single setting of this indices to change theri order and
all the code should work.

Added in options to make the Marshal's red and or black joker
wild. These are false by default, the Marshal's jokers have
initiative values following the deck. Sleeved cards when used
follow the non-wild joker and jokers (when wild) follow that.