Improve the chip tab
Made the chip tab functional.

It shows the chip pot, and the marshal's chips.
It allows the marshal to use a chip and to draw a chip from the pot.

Additionally there is now a chip allocator app that allows the marshal
to distribut chips.

Improved the assigning of chips to characters and added a report to
chat.