Add chip draw to combat tracker
Added a setting to determine whether the Marshal's deck generates a
chip for the Marshal when a black joker is drawn (defaults to off).

Modified the Draw card socket fumctions (Posse and Marshal) to
trigger a chip draw on a joker. The Marshall draw checks the
setting.

Bugfix make Chips.getColour actually return a colour. Note this
cannot distinguish between a green and temporary green, but is not
used in a context ehere this matters.