The scaffolding of editing and displaying items
The Deadlands Classic game mode will have five basic types of item

  1) The character modification. This will cover edges, hindrances and
     special features like hexes or blessings.

  2) Guns

  3) Other ranged weapons

  4) Mêlée weapons

  5) Miscellaneous items (everything else).

Each of these has different data and each of the types has their own
data model. This means that they each need a separate editor and a
separate viewer.

This change adds in the data definitions and the structure that allows
for editing and viewing each of the five types. Only the first is in
any way fleshed out at this point, the other four are placeholders.

A left click on an item in the items sidebar will open the viewer, a
right click gives a context menu with an edit option.