Split the player character into three types
The base data is slightly different in each of the Deadlands games.
Particularly in the aptitudes. Some aptitudes have concentrations in
one version but not another. Others have a different list of available
concentrations.

The differences are in the data models, which reject data they don't
recognise as valid. The sheets for editing and displaying the data
should be universal.