Improve aptitude point used and bounty used calcs
Improve the getData operations of the three player character
sheets.

The base sheet splits the character data, separating traits, aptitudes,
wound locations, chips and unclassified.

The base sheet works out if an attribute can have concentrations.
For any that can it works out if this character has any concentrations
and if there are any left that can be applied. This is added to the
context and is then available in the other two sheets that inhert from
this one.

The creation sheet works out how many aptitude points have been spent in
total and for each aptitude whether it is possible to spend more.  It
also adds the cost of the next improvement along with canImprove and
canAddConcentration booleans to each attribute in the context.

The advancement sheet works out how much bounty has been spent on each
attribute. The total spent for the character and the total available to
spend. It also adds the cost of the next improvement along with
canImprove and canAddConcentration booleans to each attribute in the
context.