Add customizable concentrations
Added in an application that allows for concentrations to be added to
any aptitude that already has concentrations in a given game world.

This is all based of the configuration file to determine which aptitudes
should be able to have concentrations.

The DataModel will reject any concentration that it doesn't recognise as
valid. When we add the here, the Data Model definition is also
dynamically determined from the configuration.