Add in the trait setting application
Add an application to set a charcter's traits from the cards it
drew when the document was created.

The application is available under a right click on the actor in
the sidebar.

The cards are available and can be dragged to traits. Using some
cards is compulsory (Jokers and Twos).  The form checks that 10
different cards have been used and that all mandatory cards have
been used. At that point it offers a submit button.

The form also has a drop down that will allocate the cards for
you, based on a character archtype.  The order that cards are
allocated has been derived from the archtypes in the core rulebook.