Add secondary tabs and adding items to characters
Adding in logic to allow document sheets to have sub-tabs, i.e. a tabbed
page that itself has tabs.

Added the ability to drag and drop items from the item browser to a
character and have them added as embedded documents. This uses
sub-tabs of the inventory tab to hold tabs for guns, melee weapons,
other ranged weapons and miscellaneous items. There are also buttons
to allow deleteion of the item and customisation of the item.

Note: A quantity field has been added to miscellaneous items, this will
allow for things like arrows and bullets. We will want to track how
many we have without cluttering the sheet with hunreds of individual
bullets. Larger things like guns can have multiple of the same thing
added, each of which can be given a specific name or a note.

Added tabs to the character create and character modify applications
that are subtabs of a new character modification tab. There are tabs for
edges, hindrances and spelllike abilities. No logic has yet been added
for these.